requirements.txt
aiogram
aiosqlite
xlsxwriter
python-dotenv
